requirements.txt
streamlit
pandas
sqlalchemy
python-dotenv
matplotlib
seaborn
scikit-learn
requests